fixed-size types should be better match for network data. remaining issue is "inf" value for acceleration, but it seems to be on both ends. suspect it is due to rendering induced lag + dynamically determined maxacceleration value.

+++ code/Network/racerpacket.h
@@ -38,10 +38,11 @@ racerPacketType_t getRacerPacketType(ENetPacket *packet);
 
 struct RPUpdateAgent {
     netObjID_t ID;
-    unsigned long a;
-    unsigned long r;
-    short w;
-    short f;
+    /* Network-safe version of SteerInfo */
+    uint32_t a;
+    uint32_t r;
+    uint32_t w;
+    uint32_t f;
 };
 
 struct RPCreateNetObj {
